Factor out transactron context wrappers
