+++ transactron/core/__init__.py
@@ -3,5 +3,6 @@ from .transaction_base import *  # noqa: F401
 from .method import *  # noqa: F401
 from .transaction import *  # noqa: F401
 from .manager import *  # noqa: F401
+from .context import *  # noqa: F401
 from .sugar import *  # noqa: F401
 from .body import *  # noqa: F401
