+++ transactron/core/manager.py
@@ -1,13 +1,10 @@
 from collections import defaultdict, deque
 from collections.abc import Callable, Iterable, Sequence, Collection, Mapping
-from typing import TypeAlias, Optional
+from typing import TypeAlias
 from os import environ
 from amaranth import *
-from amaranth.lib.wiring import Component, connect, flipped
 from itertools import chain, filterfalse, product
 import networkx
-
-from amaranth_types import AbstractComponent
 
 from transactron.utils import *
 from transactron.utils.transactron_helpers import _graph_ccs
@@ -20,7 +17,7 @@ from .method import Method
 from .tmodule import TModule
 from .schedulers import eager_deterministic_cc_scheduler
 
-__all__ = ["TransactionManager", "TransactionModule", "TransactionComponent"]
+__all__ = ["TransactionManager"]
 
 TransactionGraph: TypeAlias = Graph[TBody]
 TransactionGraphCC: TypeAlias = GraphCC[TBody]
@@ -454,100 +451,3 @@ class TransactionManager(Elaboratable):
             "transactions": {t.name: transaction_debug(t) for t in method_map.transactions},
             "methods": {m.owned_name: method_debug(m) for m in method_map.methods},
         }
-
-
-class TransactionModule(Elaboratable):
-    """
-    `TransactionModule` is used as wrapper on `Elaboratable` classes,
-    which adds support for transactions. It creates a
-    `TransactionManager` which will handle transaction scheduling
-    and can be used in definition of `Method`\\s and `Transaction`\\s.
-    The `TransactionManager` is stored in a `DependencyManager`.
-    """
-
-    def __init__(
-        self,
-        elaboratable: HasElaborate,
-        dependency_manager: Optional[DependencyManager] = None,
-        transaction_manager: Optional[TransactionManager] = None,
-    ):
-        """
-        Parameters
-        ----------
-        elaboratable: HasElaborate
-            The `Elaboratable` which should be wrapped to add support for
-            transactions and methods.
-        dependency_manager: DependencyManager, optional
-            The `DependencyManager` to use inside the transaction module.
-            If omitted, a new one is created.
-        transaction_manager: TransactionManager, optional
-            The `TransactionManager` to use inside the transaction module.
-            If omitted, a new one is created.
-        """
-        if transaction_manager is None:
-            transaction_manager = TransactionManager()
-        if dependency_manager is None:
-            dependency_manager = DependencyManager()
-        self.manager = dependency_manager
-        self.manager.add_dependency(TransactionManagerKey(), transaction_manager)
-        self.elaboratable = elaboratable
-
-    def context(self) -> DependencyContext:
-        return DependencyContext(self.manager)
-
-    def elaborate(self, platform):
-        with silence_mustuse(self.manager.get_dependency(TransactionManagerKey())):
-            with self.context():
-                elaboratable = Fragment.get(self.elaboratable, platform)
-
-        m = Module()
-
-        m.submodules.main_module = elaboratable
-        m.submodules.transactionManager = self.transaction_manager = self.manager.get_dependency(
-            TransactionManagerKey()
-        )
-
-        return m
-
-
-class TransactionComponent(TransactionModule, Component):
-    """Top-level component for Transactron projects.
-
-    The `TransactronComponent` is a wrapper on `Component` classes,
-    which adds Transactron support for the wrapped class. The use
-    case is to wrap a top-level module of the project, and pass the
-    wrapped module for simulation, HDL generation or synthesis.
-    The ports of the wrapped component are forwarded to the wrapper.
-
-    It extends the functionality of `TransactionModule`.
-    """
-
-    def __init__(
-        self,
-        component: AbstractComponent,
-        dependency_manager: Optional[DependencyManager] = None,
-        transaction_manager: Optional[TransactionManager] = None,
-    ):
-        """
-        Parameters
-        ----------
-        component: Component
-            The `Component` which should be wrapped to add support for
-            transactions and methods.
-        dependency_manager: DependencyManager, optional
-            The `DependencyManager` to use inside the transaction component.
-            If omitted, a new one is created.
-        transaction_manager: TransactionManager, optional
-            The `TransactionManager` to use inside the transaction component.
-            If omitted, a new one is created.
-        """
-        TransactionModule.__init__(self, component, dependency_manager, transaction_manager)
-        Component.__init__(self, component.signature)
-
-    def elaborate(self, platform):
-        m = super().elaborate(platform)
-
-        assert isinstance(self.elaboratable, Component)  # for typing
-        connect(m, flipped(self), self.elaboratable)
-
-        return m
